% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deg_days}
\alias{deg_days}
\title{Degree Days}
\format{dataframe with 7,812 rows and 3 variables
\describe{
  \item{date}{calendar date, monthly value 1979-2000}
  \item{watershed}{CVPIA watershed name}
  \item{degdays}{monthly ATU in °C}
}}
\source{
\itemize{
  \item HEC-5Q Model Output and Double Regression Temperature Modeling: Michael Wright \email{mwright@usbr.gov}
  \item Data Wrangling and Additional Temperature Modeling: Sadie Gill  \email{sgill@flowwest.com}
}
}
\usage{
deg_days
}
\description{
The monthly accumulated degree days (celsius)
}
\details{
For watersheds with HEC-5Q modeled results, the calculation for degree days
is the sum of the daily mean values for each month.

For the other regions, the calculation for degree days is the monthly mean
temperature multiplied by the number of days in the month.

For more details about the temperature modeling see \code{\link{juv_temp}}
}
\keyword{datasets}
