% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delta_temps}
\alias{delta_temps}
\title{Delta Temperature}
\format{dataframe with 504 rows and 3 variables
\describe{
  \item{date}{calendar date, monthly value 1980-1999}
  \item{watershed}{CVPIA watershed name}
  \item{monthly_mean_temp_c}{modeled monthly mean temperature in °C}
}}
\source{
Data Wrangling and Additional Temperature Modeling: Sadie Gill  \email{sgill@flowwest.com}
}
\usage{
delta_temps
}
\description{
Rearing temperature in the North and South Delta
}
\details{
The deltas' water temperature was modeled as a function of air temperature.

North Delta was modeled using a sample of recent measured water temperature at
CDEC station \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=EMM}{EMM}
and NOAA CDO air temperature data from \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00040232/detail}{Antioch}.

South Delta was modeled using a sample of recent measured water temperature at
CDEC station \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=MTB}{MTB}
and NOAA CDO air temperature data \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023237/detail}{Stockton Airport}.

Air Temperature Data Source:
 \itemize{
   \item NOAA Climate Data Online \href{https://www.ncdc.noaa.gov/cdo-web/}{(CDO)} accessed using the \href{https://github.com/ropensci/rnoaa}{\code{rnoaa}} R package developed by \href{https://ropensci.org/}{rOpenSci}.
   }

 Stream Water Temperature Data Sources:
 \itemize{
   \item DWR California Data Exchange Center \href{http://cdec.water.ca.gov/index.html}{(CDEC)} accessed using the \href{https://github.com/flowwest/cdecretrieve}{\code{CDECRetrieve}} R packaged developed by \href{http://www.flowwest.com/}{FlowWest}.
 }
}
\keyword{datasets}
