% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_mean_temperature}
\alias{monthly_mean_temperature}
\title{Monthly Mean Water Temperature}
\format{dataframe with 240 rows and 32 variables
\describe{
 \item{date}{calendar date}
 \item{Upper Sacramento River}{Hec 5Q}
 \item{Antelope Creek}{Cow Creek temperatue used as surrogate}
 \item{Battle Creek}{Hec 5Q}
 \item{Bear Creek}{Cow Creek temperatue used as surrogate}
 \item{Big Chico Creek}{Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=BIC}{BIC} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00046685/detail}{USC00046685}}
 \item{Butte Creek}{Estimated mean monthly water temperature from a linear model fitted with water temperature data from CDEC Gage ID: \href{http://cdec.water.ca.gov/cgi-progs/staMeta?station_id=BCD}{BCD} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00046685/detail}{USC00046685}}
 \item{Clear Creek}{Hec 5Q}
 \item{Cottonwood Creek}{Hec 5Q}
 \item{Cow Creek}{Hec 5Q}
 \item{Deer Creek}{Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11383500}{11383500} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00041715/detail}{USC00041715}}
 \item{Elder Creek}{Thomes Creek temperatue used as surrogate}
 \item{Mill Creek}{Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11381500}{11381500} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024216/detail}{USW00024216}}
 \item{Paynes Creek}{Cow Creek temperatue used as surrogate}
 \item{Stony Creek}{Hec 5Q}
 \item{Thomes Creek}{Hec 5Q}
 \item{Upper-mid Sacramento River}{Hec 5Q}
 \item{Sutter Bypass}{TBD}
 \item{Bear River}{TBD}
 \item{Feather River}{American River temperatue used as surrogate}
 \item{Yuba River}{Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11421000}{11421000} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00024216/detail}{USW00024216}}
 \item{Lower-mid Sacramento River}{Hec 5Q}
 \item{Yolo Bypass}{TBD}
 \item{American River}{Hec 5Q}
 \item{Lower Sacramento River}{Measured mean monthly water temperature from  USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11447650}{11447650} with imputed missing values using \href{https://www.rdocumentation.org/packages/forecast/versions/8.1/topics/na.interp}{\code{forecast::na.interp}}}
 \item{Calaveras River}{Mokelumne River temperatue used as surrogate}
 \item{Cosumnes River}{Estimated mean monthly water temperature from a linear model fitted with water temperature data from USGS Gage ID: \href{https://waterdata.usgs.gov/nwis/inventory?agency_code=USGS&site_no=11335000}{11335000} and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USW00023271/detail}{USW00023271}}
 \item{Mokelumne River}{Estimated mean monthly water temperature from a linear model fitted with water temperature data provided by EBMUD measured near Victor, CA and air temperature data from NOAA CDO Station Id: \href{https://www.ncdc.noaa.gov/cdo-web/datasets/GSOM/stations/GHCND:USC00045032/detail}{USC00045032}}
 \item{Merced River}{Hec 5Q}
 \item{Stanislaus River}{Hec 5Q}
 \item{Tuolumne River}{Hec 5Q}
 \item{San Joaquin River}{Hec 5Q}
 }}
\source{
\itemize{
  \item Hec 5Q Model Output: Michael Wright \email{mwright@usbr.gov}
  \item Data Wrangling and Additional Temperature Modeling: Sadie Gill  \email{sgill@flowwest.com}
  \item Temperature Surrogate Selection: Mark Tompkins \email{mtompkins@flowwest.com} and Mike Urkov \email{mike.urkov@gmail.com}
}
}
\usage{
monthly_mean_temperature
}
\description{
A dataset containing the monthly mean water temperature in
celsius for each CVPIA watershed 1980-1999.
}
\details{
Combo of Mike W, subs, and water ~ air modeling. data sources for modeling:
NOAA, CDEC, USGS
}
\keyword{datasets}
