% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stream_temperature}
\alias{stream_temperature}
\title{Rearing Stream Temperature}
\format{
a 3 dimensional array [31 watersheds, 12 months, 21 years]
}
\usage{
stream_temperature
}
\description{
The 1979-1999 rearing temperature in °C
}
\details{
Calculated using \code{\link{monthly_mean_temperature}}
}
\keyword{datasets}
